pyproject.toml
[tool.poetry]
name = "ragx"
version = "0.3.0"
description = ""
authors = ["gaojian <olivetree123@163.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.8.1"
build = "1.2.2"
cachecontrol = "0.14.0"
certifi = "2024.8.30"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
colorama = "0.4.6"
crashtest = "0.4.1"
distlib = "0.3.8"
django = "5.1.2"
django-ninja = "1.3.0"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.16.1"
gunicorn = "^23.0.0"
idna = "3.10"
importlib-metadata = "8.5.0"
installer = "0.7.0"
jaraco-classes = "3.4.0"
keyring = "24.3.1"
more-itertools = "10.5.0"
msgpack = "1.1.0"
mysqlclient = "2.2.4"
packaging = "24.1"
pexpect = "4.9.0"
pkginfo = "1.11.1"
platformdirs = "4.3.6"
ptyprocess = "0.7.0"
pyproject-hooks = "1.2.0"
pywin32-ctypes = "0.2.3"
rapidfuzz = "3.10.0"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
sqlparse = "0.5.1"
tomli = "2.0.1"
tomlkit = "0.13.2"
trove-classifiers = "2024.9.12"
typing-extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.3"
virtualenv = "20.26.6"
zipp = "3.20.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pytomlenv = "^0.2.0"
celery = "^5.4.0"
redis = "^5.2.0"
langchain = "^0.3.7"
pymilvus = {extras = ["model"], version = "^2.4.9"}
colorlog = "^6.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
